# services/rest/docs/openapi-documentation.yml
openapi: 3.1.0
info:
  title: Movie API
  version: 1.0.0
  description: REST API for authentication, movies, directors, and production companies.
servers:
  - url: http://localhost:5286
paths:
  /api/v1/Auth/login:
    post:
      summary: Authenticate a user
      requestBody:
        description: Credentials used to obtain an Auth0 token
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseError'
        "500":
          description: Internal server error
  /api/v1/Auth/logout/{jwtId}:
    post:
      summary: Logout and blacklist the current token
      parameters:
        - name: jwtId
          in: path
          required: true
          description: JWT ID (jti) of the token to blacklist
          schema:
            type: string
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseError'
        "401":
          description: Unauthorized or malformed token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseError'
        "403":
          description: JWT ID in path does not match the authenticated token
        "500":
          description: Internal server error
...
  /api/v1/Directors:
    get:
      summary: Get all directors (paginated)
      parameters:
        - name: page
          in: query
          required: false
          description: Page number (starts at 1)
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
            default: 1
        - name: pageSize
          in: query
          required: false
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
            default: 10
        - name: search
          in: query
          required: false
          description: Search term for directors
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Paginated list of directors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDirectorResult'
        "204":
          description: No directors found
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /api/v1/Directors/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Director ID
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
    get:
      summary: Get a director by ID
      responses:
        "200":
          description: Director found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorHateoasDto'
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Unauthorized
        "404":
          description: Director not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseMessage'
        "500":
          description: Internal server error
...
components:
  schemas:
    ErrorResponseBadRequest:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        traceId:
          type: string
      example:
        type: https://tools.ietf.org/html/rfc9110#section-15.5.1
        title: One or more validation errors occurred.
        status: 400
        errors:
          Password:
            - The Password field is required.
        traceId: 00-d1bd44b0408fa52dfb31a0918a49bb51
    ErrorResponseMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ErrorResponseError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    AuthTokenResponse:
      type: object
      description: Auth0 token response payload
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        id_token:
          type: string
    LogoutResponse:
      type: object
      properties:
        message:
          type: string
...
    Link:
      type: object
      required:
        - href
        - rel
        - method
      properties:
        href:
          type: string
        rel:
          type: string
        method:
          type: string
        templated:
          type: boolean
...
    DirectorHateoasDto:
      type: object
      required:
        - id
        - firstName
        - lastName
        - _links
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...
    PaginatedDirectorResult:
      type: object
      required:
        - _embedded
        - page
        - pageSize
        - totalCount
        - totalPages
        - _links
      properties:
        _embedded:
          type: object
          properties:
            directors:
              type: array
              items:
                $ref: '#/components/schemas/DirectorHateoasDto'
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer 
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...